Bug 3 Replication Complete
Bug 3 has been Replicated and an FAT to prove it has been provided.

As a note, the reason this commit remains in the same branch as Bug 2 is that aside for bug 2 being a non-issue , IntelliJ apparently takes issue with branches changing when not using its own internal git client (which i didnt know existed) so this branch will be used for this and any future bugs.